+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.0.3] - 7 September 2022
+
+- adjust default `SkadiResponsive` value breakpoint
+- Fix `SkadiResponsive.isMobile` return wrong value
+
 ## [0.0.2] - 29 August 2022
 
 - Add more widgets